\name{rWBData-package}
\alias{rWBData-package}
\alias{rWBData}
\docType{package}
\title{
Access World Bank's Open Data in R
}
\description{
Convert World Bank's Open Data to a data frame or download bulk data (in format of .csv, .xlsx, .zip ...) directly.
}
\details{
\tabular{ll}{
Package: \tab rWBData\cr
Type: \tab Package\cr
Version: \tab 0.1.1\cr
Date: \tab 2014-08-21\cr
License: \tab GPL-3\cr
}
}
\author{
Ming-Jer Lee

Maintainer: Ming-Jer Lee <mingjerli@gmail.com>
}
