\name{updateCategoryList}
\alias{updateCategoryList}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
  Update database category
}
\description{
  This function will update the database category from world bank open data (http://data.worldbank.org/).
}
\usage{
updateCategoryList()
}
\author{
Ming-Jer Lee <mingjerli@gmail.com>
}
\examples{
updateCategoryList()
}
